+++ test/rule_test.dart
@@ -192,4 +192,22 @@ enum Direction { left, right }
       ],
     );
   }
+
+  /// https://github.com/huanghui1998hhh/prefer_shorthands/issues/11
+  void test_11() async {
+    await assertDiagnostics(
+      '''
+final object = <String, Direction>{}..['a'] = Direction.left;
+final object2 = <String, dynamic>{}..['a'] = Direction.left;
+
+void main() {
+  object['a'] = Direction.left;
+  object2['a'] = Direction.left;
+}
+
+enum Direction { left, right }
+''',
+      [lint(46, 14), lint(154, 14)],
+    );
+  }
 }